demo/requirements.txt
streamlit
torch
torchvision
Pillow
ultralytics
pytorch-lightning